Rename interval parameter to delay in RetryNode and update related tests
