+++ tests/workflows/basic_retry_node_delay/testing/test_workflow.py
@@ -1,10 +1,10 @@
 from datetime import datetime, timedelta
 
-from tests.workflows.basic_retry_node_interval.workflow import SimpleRetryExample
+from tests.workflows.basic_retry_node_delay.workflow import SimpleRetryExample
 
 
 def test_run_workflow__happy_path():
-    # GIVEN a workflow that references a Retry example with interval
+    # GIVEN a workflow that references a Retry example with delay
     workflow = SimpleRetryExample()
 
     # WHEN the workflow is run
@@ -17,5 +17,5 @@ def test_run_workflow__happy_path():
     # AND the output should match the environment variable
     assert terminal_event.outputs == {"final_value": 3}
 
-    # AND the duration should be (retry - 1) * interval
+    # AND the duration should be (retry - 1) * delay
     assert terminal_event.timestamp - datetime_start > timedelta(seconds=0.2)
